test(useEffect): Test pass useIsAdmin
