+++ components/routes/admin-tech-table.tsx
@@ -1,6 +1,6 @@
 "use client"
 
-import React, { useEffect, useState } from "react";
+import React, { useState } from "react";
 import { Table, TableHeader, TableColumn, TableBody, TableRow, TableCell, Pagination, getKeyValue, Chip, Tooltip, User } from "@nextui-org/react";
 import { IJsonTech } from "@/types";
 import { CiEdit } from "react-icons/ci";
@@ -9,23 +9,14 @@ import { lenguajesResources } from "@/data/data";
 import UserDescAdminTechTable from "./user-desc-admin-tech-table";
 import TopContentAdminTechTable from "./top-content-admin-tech-table";
 import DeleteTechButton from "./delete-admin-tech-button";
-import { FlattenedAdmin, useIsAdmin } from "@/utils/isAdmin";
-import { useActiveAccount } from "thirdweb/react";
-import { fetchAdmins } from "@/data/fetch";
+import { FlattenedAdmin } from "@/utils/isAdmin";
+import useIsAdmin from "@/hooks/useIsAdmin";
 
 
 interface AdminTechTableProps {
   lenguajes: IJsonTech[];
   admins: FlattenedAdmin[];
 }
-// export const useIsAdmin = (account:any) => {
-//   // const account = useActiveAccount();
-  
-
-  
-
-//   return isAdmin;
-// };
 
 
 // El revalidate del delete se hace en su funcion del servidor, pero el revalidate del update y del create se llama desde el componente del cliente
@@ -33,40 +24,39 @@ const AdminTechTable: React.FC<AdminTechTableProps> = ({lenguajes, admins}) => {
   const [page, setPage] = useState<number>(1);
 
   const [error, setError] = useState<string | null>(null);
-  // const [showSpinner, setShowSpinner] = useState<boolean>(false);
 
   const rowsPerPage = 4;
 
   // is Admin - [ ] Ver si funciona sin el useEffect -> NO ES NECESARIO(Solo hay que llamar siempre a la re-asignación)
-  const account = useActiveAccount();
-    const [isAdmin, setIsAdmin] = useState(false);
+  // const account = useActiveAccount();
+  //   const [isAdmin, setIsAdmin] = useState(false);
 
-    useEffect(() => {
-      console.log("admins admin techTable: ", admins)
-        const checkIsAdmin = async () => {
-            try {
-                // const admins = await fetchAdmins(); // Obtener la lista de administradores
-                if (account?.address) {
-                    // Verificar si alguna dirección de admin coincide con la dirección de la cuenta activa
-                    const isAdminUser = admins.some(admin => admin.address === account.address);
-                    setIsAdmin(isAdminUser);
-                    console.log("isAdmin (TechTable): ", isAdminUser);
-                    console.log("address: ", account.address);
-                }
-            } catch (error) {
-                console.error('Error al obtener la lista de administradores', error);
-                // Manejar el error según sea necesario
-            }
-        };
+  //   useEffect(() => {
+  //     console.log("admins admin techTable: ", admins)
+  //       const checkIsAdmin = async () => {
+  //           try {
+  //               // const admins = await fetchAdmins(); // Obtener la lista de administradores
+  //               if (account?.address) {
+  //                   // Verificar si alguna dirección de admin coincide con la dirección de la cuenta activa
+  //                   const isAdminUser = admins.some(admin => admin.address === account.address);
+  //                   setIsAdmin(isAdminUser);
+  //                   console.log("isAdmin (TechTable): ", isAdminUser);
+  //                   console.log("address: ", account.address);
+  //               }
+  //           } catch (error) {
+  //               console.error('Error al obtener la lista de administradores', error);
+  //               // Manejar el error según sea necesario
+  //           }
+  //       };
 
-        checkIsAdmin(); // Llamar a la función para verificar si la cuenta es administrador
+  //       checkIsAdmin(); // Llamar a la función para verificar si la cuenta es administrador
 
-    }, [account]);
+  //   }, [account]);
   // const account = useActiveAccount();
   
   
   // const isAdmin = account?.address === "0x490bb233c707A0841cA52979Be4D88B6621d1988";
-  // const { isAdmin, account } = await useIsAdmin();
+  const { isAdmin, account } = useIsAdmin(admins);
   
 
   // Prepare pagination
